@@ -4,8 +4,10 @@
   "private": true,
   "homepage": "https://peckeby.github.io/noname-digital/",
   "dependencies": {
-    "@material-ui/core": "^4.12.4",
-    "@material-ui/icons": "^4.11.3",
+    "@emotion/react": "^11.10.6",
+    "@emotion/styled": "^11.10.6",
+    "@mui/icons-material": "^5.11.11",
+    "@mui/material": "^5.11.11",
     "@reduxjs/toolkit": "^1.9.3",
     "@testing-library/jest-dom": "^5.16.3",
     "@testing-library/react": "^12.1.4",
@@ -13,18 +15,17 @@
     "firebase": "^9.17.2",
     "install": "^0.13.0",
     "npm": "^9.6.0",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-redux": "^8.0.5",
     "react-redux-firebase": "^3.11.0",
     "react-router-dom": "^6.8.2",
-    "react-scripts": "^4.0.1",
+    "react-scripts": "^5.0.1",
     "redux": "^4.2.1",
     "redux-firestore": "^1.0.0",
     "redux-persist": "^6.0.0",
     "router-dom": "^2.2.10",
     "sass": "^1.58.3",
-    "styled-components": "^5.3.6",
     "web-vitals": "^2.1.3"
   },
   "scripts": {
