@@ -1,16 +1,14 @@
 import React from 'react';
-import {
-  Button,
-  CardActions,
-  CardContent,
-  CardMedia,
-  Typography,
-} from '@material-ui/core';
-import Card from '@material-ui/core/Card';
+import Card from '@mui/material/Card';
+import CardActions from '@mui/material/CardActions';
+import CardContent from '@mui/material/CardContent';
+import CardMedia from '@mui/material/CardMedia';
+import Button from '@mui/material/Button';
+import Typography from '@mui/material/Typography';
 
 export default function ProductCards({ product }) {
   return (
-    <Card>
+    <Card sx={{ width: 400 }} key={product.id}>
       <CardMedia
         component="img"
         alt={product.name}
